+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 ###Changelog
 ####next release
 * Pebble: Huge speedup for app/firmware installation.
-* Pebble: Use a separate notification for installation procedure
+* Pebble: Use a separate notification with progress bar for installation procedure
 * Pebble: Bugfix for beeing stuck while waiting for a slot, when none is available
 * Mi Band: Display connection status in notification (previously Pebble only)